+++ src/test/java/step4/domain/CarsTest.java
@@ -2,7 +2,7 @@ package step4.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import step4.RecordDto;
+import step4.dto.RecordDto;
 
 import java.util.ArrayList;
 import java.util.Arrays;
